Remove dependencies utils.js, das.js, and color.js
