--- color.js
@@ -1,126 +1,0 @@
-/* -*- mode: javascript; c-basic-offset: 4; indent-tabs-mode: nil -*- */
-
-// 
-// Dalliance Genome Explorer
-// (c) Thomas Down 2006-2010
-//
-// color.js
-//
-
-define(function(require, exports, module) {
-"use strict";
-
-function DColour(red, green, blue, name) {
-    this.red = red|0;
-    this.green = green|0;
-    this.blue = blue|0;
-    if (name) {
-        this.name = name;
-    }
-}
-
-DColour.prototype.toSvgString = function() {
-    if (!this.name) {
-        this.name = "rgb(" + this.red + "," + this.green + "," + this.blue + ")";
-    }
-
-    return this.name;
-}
-
-function hex2(x) {
-    var y = '00' + x.toString(16);
-    return y.substring(y.length - 2);
-}
-
-DColour.prototype.toHexString = function() {
-    return '#' + hex2(this.red) + hex2(this.green) + hex2(this.blue);
-}
-
-var palette = {
-    red: new DColour(255, 0, 0, 'red'),
-    green: new DColour(0, 255, 0, 'green'),
-    blue: new DColour(0, 0, 255, 'blue'),
-    yellow: new DColour(255, 255, 0, 'yellow'),
-    white: new DColour(255, 255, 255, 'white'),
-    black: new DColour(0, 0, 0, 'black'),
-    gray: new DColour(180, 180, 180, 'gray'),
-    grey: new DColour(180, 180, 180, 'grey'),
-    lightskyblue: new DColour(135, 206, 250, 'lightskyblue'),
-    lightsalmon: new DColour(255, 160, 122, 'lightsalmon'),
-    hotpink: new DColour(255, 105, 180, 'hotpink')
-};
-
-var COLOR_RE = new RegExp('^#([0-9A-Fa-f]{2})([0-9A-Fa-f]{2})([0-9A-Fa-f]{2})$');
-var CSS_COLOR_RE = /rgb\(([0-9]+),([0-9]+),([0-9]+)\)/
-
-function dasColourForName(name) {
-    var c = palette[name];
-    if (!c) {
-        var match = COLOR_RE.exec(name);
-        if (match) {
-            c = new DColour(('0x' + match[1])|0, ('0x' + match[2])|0, ('0x' + match[3])|0, name);
-            palette[name] = c;
-        } else {
-    	    match = CSS_COLOR_RE.exec(name);
-    	    if (match) {
-        		c = new DColour(match[1]|0, match[2]|0, match[3]|0, name);
-        		palette[name] = c;
-	       } else {
-		      console.log("couldn't handle color: " + name);
-		      c = palette.black;
-		      palette[name] = c;
-	       }
-        }
-    }
-    return c;
-}
-
-function makeColourSteps(steps, stops, colours) {
-    var dcolours = [];
-    for (var ci = 0; ci < colours.length; ++ci) {
-        dcolours.push(dasColourForName(colours[ci]));
-    }
-
-    var grad = [];
-  STEP_LOOP:
-    for (var si = 0; si < steps; ++si) {
-        var rs = (1.0 * si) / (steps-1);
-        var score = stops[0] + (stops[stops.length -1] - stops[0]) * rs;
-        for (var i = 0; i < stops.length - 1; ++i) {
-            if (score >= stops[i] && score <= stops[i+1]) {
-                var frac = (score - stops[i]) / (stops[i+1] - stops[i]);
-                var ca = dcolours[i];
-                var cb = dcolours[i+1];
-
-                var fill = new DColour(
-                    ((ca.red * (1.0 - frac)) + (cb.red * frac))|0,
-                    ((ca.green * (1.0 - frac)) + (cb.green * frac))|0,
-                    ((ca.blue * (1.0 - frac)) + (cb.blue * frac))|0
-                ).toSvgString();
-                grad.push(fill);
-
-                continue STEP_LOOP;
-            }
-        }
-        throw 'Bad step';
-    }
-
-    return grad;
-}
-
-function makeGradient(steps, color1, color2, color3) {
-    if (color3) {
-        return makeColourSteps(steps, [0, 0.5, 1], [color1, color2, color3]);
-    } else {
-        return makeColourSteps(steps, [0, 1], [color1, color2]);
-    }
-}
-
-if (typeof(module) !== 'undefined') {
-    module.exports = {
-        makeColourSteps: makeColourSteps,
-        makeGradient: makeGradient,
-        dasColourForName: dasColourForName
-    };
-}
-});
