+++ index.html
@@ -19,7 +19,7 @@
                     bb = _bb;
                     err = _err;
                     cb = true;
-                    
+
                     $('body').append("<h3>Information about test.bigwig" + "</h3>");
                     $('body').append("Version: " + bb.version + "<br>");
                     $('body').append("Number of Zoom Levels: " + bb.numZoomLevels + "<br>");
