--- utils.js
@@ -1,497 +1,0 @@
-/* -*- mode: javascript; c-basic-offset: 4; indent-tabs-mode: nil -*- */
-
-// 
-// Dalliance Genome Explorer
-// (c) Thomas Down 2006-2010
-//
-// utils.js: odds, sods, and ends.
-//
-
-define(function(require, exports, module) {
-"use strict";
-
-if (typeof(require) !== 'undefined') {
-    var sha1 = require('./sha1');
-    var b64_sha1 = sha1.b64_sha1;
-}
-
-var NUM_REGEXP = new RegExp('[0-9]+');
-
-function stringToNumbersArray(str) {
-    var nums = new Array();
-    var m;
-    while (m = NUM_REGEXP.exec(str)) {
-        nums.push(m[0]);
-        str=str.substring(m.index + (m[0].length));
-    }
-    return nums;
-}
-
-var STRICT_NUM_REGEXP = new RegExp('^[0-9]+$');
-
-function stringToInt(str) {
-    str = str.replace(new RegExp(',', 'g'), '');
-    if (!STRICT_NUM_REGEXP.test(str)) {
-        return null;
-    }
-    return str|0;
-}
-
-function pushnew(a, v) {
-    for (var i = 0; i < a.length; ++i) {
-        if (a[i] == v) {
-            return;
-        }
-    }
-    a.push(v);
-}
-
-function pusho(obj, k, v) {
-    if (obj[k]) {
-        obj[k].push(v);
-    } else {
-        obj[k] = [v];
-    }
-}
-
-function pushnewo(obj, k, v) {
-    var a = obj[k];
-    if (a) {
-        for (var i = 0; i < a.length; ++i) {    // indexOf requires JS16 :-(.
-            if (a[i] == v) {
-                return;
-            }
-        }
-        a.push(v);
-    } else {
-        obj[k] = [v];
-    }
-}
-
-
-function pick(a, b, c, d)
-{
-    if (a) {
-        return a;
-    } else if (b) {
-        return b;
-    } else if (c) {
-        return c;
-    } else if (d) {
-        return d;
-    }
-}
-
-function pushnew(l, o)
-{
-    for (var i = 0; i < l.length; ++i) {
-        if (l[i] == o) {
-            return;
-        }
-    }
-    l.push(o);
-}
-
-
-
-function arrayIndexOf(a, x) {
-    if (!a) {
-        return -1;
-    }
-
-    for (var i = 0; i < a.length; ++i) {
-        if (a[i] === x) {
-            return i;
-        }
-    }
-    return -1;
-}
-
-function arrayRemove(a, x) {
-    var i = arrayIndexOf(a, x);
-    if (i >= 0) {
-        a.splice(i, 1);
-        return true;
-    }
-    return false;
-}
-
-//
-// DOM utilities
-//
-
-
-function makeElement(tag, children, attribs, styles)
-{
-    var ele = document.createElement(tag);
-    if (children) {
-        if (! (children instanceof Array)) {
-            children = [children];
-        }
-        for (var i = 0; i < children.length; ++i) {
-            var c = children[i];
-            if (c) {
-                if (typeof c == 'string') {
-                    c = document.createTextNode(c);
-                } else if (typeof c == 'number') {
-                    c = document.createTextNode('' + c);
-                }
-                ele.appendChild(c);
-            }
-        }
-    }
-    
-    if (attribs) {
-        for (var l in attribs) {
-            try {
-                ele[l] = attribs[l];
-            } catch (e) {
-                console.log('error setting ' + l);
-                throw(e);
-            }
-        }
-    }
-    if (styles) {
-        for (var l in styles) {
-            ele.style[l] = styles[l];
-        }
-    }
-    return ele;
-}
-
-function makeElementNS(namespace, tag, children, attribs)
-{
-    var ele = document.createElementNS(namespace, tag);
-    if (children) {
-        if (! (children instanceof Array)) {
-            children = [children];
-        }
-        for (var i = 0; i < children.length; ++i) {
-            var c = children[i];
-            if (typeof c == 'string') {
-                c = document.createTextNode(c);
-            }
-            ele.appendChild(c);
-        }
-    }
-    
-    setAttrs(ele, attribs);
-    return ele;
-}
-
-var attr_name_cache = {};
-
-function setAttr(node, key, value)
-{
-    var attr = attr_name_cache[key];
-    if (!attr) {
-        var _attr = '';
-        for (var c = 0; c < key.length; ++c) {
-            var cc = key.substring(c, c+1);
-            var lcc = cc.toLowerCase();
-            if (lcc != cc) {
-                _attr = _attr + '-' + lcc;
-            } else {
-                _attr = _attr + cc;
-            }
-        }
-        attr_name_cache[key] = _attr;
-        attr = _attr;
-    }
-    node.setAttribute(attr, value);
-}
-
-function setAttrs(node, attribs)
-{
-    if (attribs) {
-        for (var l in attribs) {
-            setAttr(node, l, attribs[l]);
-        }
-    }
-}
-
-
-
-function removeChildren(node)
-{
-    if (!node || !node.childNodes) {
-        return;
-    }
-
-    while (node.childNodes.length > 0) {
-        node.removeChild(node.firstChild);
-    }
-}
-
-
-
-//
-// WARNING: not for general use!
-//
-
-function miniJSONify(o, exc) {
-    if (typeof o === 'undefined') {
-        return 'undefined';
-    } else if (o == null) {
-        return 'null';
-    } else if (typeof o == 'string') {
-        return "'" + o + "'";
-    } else if (typeof o == 'number') {
-        return "" + o;
-    } else if (typeof o == 'boolean') {
-        return "" + o;
-    } else if (typeof o == 'object') {
-        if (o instanceof Array) {
-            var s = null;
-            for (var i = 0; i < o.length; ++i) {
-                s = (s == null ? '' : (s + ', ')) + miniJSONify(o[i], exc);
-            }
-            return '[' + (s?s:'') + ']';
-        } else {
-            exc = exc || {};
-            var s = null;
-            for (var k in o) {
-                if (exc[k])
-                    continue;
-                if (k != undefined && typeof(o[k]) != 'function') {
-                    s = (s == null ? '' : (s + ', ')) + k + ': ' + miniJSONify(o[k], exc);
-                }
-            }
-            return '{' + (s?s:'') + '}';
-        }
-    } else {
-        return (typeof o);
-    }
-}
-
-function shallowCopy(o) {
-    var n = {};
-    for (var k in o) {
-        n[k] = o[k];
-    }
-    return n;
-}
-
-function Observed(x) {
-    this.value = x;
-    this.listeners = [];
-}
-
-Observed.prototype.addListener = function(f) {
-    this.listeners.push(f);
-}
-
-Observed.prototype.addListenerAndFire = function(f) {
-    this.listeners.push(f);
-    f(this.value);
-}
-
-Observed.prototype.removeListener = function(f) {
-    arrayRemove(this.listeners, f);
-}
-
-Observed.prototype.get = function() {
-    return this.value;
-}
-
-Observed.prototype.set = function(x) {
-    this.value = x;
-    for (var i = 0; i < this.listeners.length; ++i) {
-        this.listeners[i](x);
-    }
-}
-
-function Awaited() {
-    this.queue = [];
-}
-
-Awaited.prototype.provide = function(x) {
-    if (this.res !== undefined) {
-        throw "Resource has already been provided.";
-    }
-
-    this.res = x;
-    for (var i = 0; i < this.queue.length; ++i) {
-        this.queue[i](x);
-    }
-    this.queue = null;   // avoid leaking closures.
-}
-
-Awaited.prototype.await = function(f) {
-    if (this.res !== undefined) {
-        f(this.res);
-        return this.res;
-    } else {
-        this.queue.push(f);
-    }
-}
-
-var __dalliance_saltSeed = 0;
-
-function saltURL(url) {
-    return url + '?salt=' + b64_sha1('' + Date.now() + ',' + (++__dalliance_saltSeed));
-}
-
-function textXHR(url, callback, opts) {
-    if (opts && opts.salt) 
-        url = saltURL(url);
-
-    try {
-        var timeout;
-        if (opts.timeout) {
-            timeout = setTimeout(
-                function() {
-                    console.log('timing out ' + url);
-                    req.abort();
-                    return callback(null, 'Timeout');
-                },
-                opts.timeout
-            );
-        }
-
-        var req = new XMLHttpRequest();
-        req.onreadystatechange = function() {
-    	    if (req.readyState == 4) {
-                if (timeout)
-                    clearTimeout(timeout);
-    	        if (req.status < 200 || req.status >= 300) {
-    		    callback(null, 'Error code ' + req.status);
-    	        } else {
-    		    callback(req.responseText);
-    	        }
-    	    }
-        };
-        
-        req.open('GET', url, true);
-        req.responseType = 'text';
-
-        if (opts && opts.credentials) {
-            req.withCredentials = true;
-        }
-        req.send('');
-    } catch (e) {
-        callback(null, 'Exception ' + e);
-    }
-}
-
-function relativeURL(base, rel) {
-    // FIXME quite naive -- good enough for trackhubs?
-
-    if (rel.indexOf('http:') == 0 || rel.indexOf('https:') == 0) {
-        return rel;
-    }
-
-    var li = base.lastIndexOf('/');
-    if (li >= 0) {
-        return base.substr(0, li + 1) + rel;
-    } else {
-        return rel;
-    }
-}
-
-var AMINO_ACID_TRANSLATION = {
-    'TTT': 'F',
-    'TTC': 'F',
-    'TTA': 'L',
-    'TTG': 'L',
-    'CTT': 'L',
-    'CTC': 'L',
-    'CTA': 'L',
-    'CTG': 'L',
-    'ATT': 'I',
-    'ATC': 'I',
-    'ATA': 'I',
-    'ATG': 'M',
-    'GTT': 'V',
-    'GTC': 'V',
-    'GTA': 'V',
-    'GTG': 'V',
-    'TCT': 'S',
-    'TCC': 'S',
-    'TCA': 'S',
-    'TCG': 'S',
-    'CCT': 'P',
-    'CCC': 'P',
-    'CCA': 'P',
-    'CCG': 'P',
-    'ACT': 'T',
-    'ACC': 'T',
-    'ACA': 'T',
-    'ACG': 'T',
-    'GCT': 'A',
-    'GCC': 'A',
-    'GCA': 'A',
-    'GCG': 'A',
-    'TAT': 'Y',
-    'TAC': 'Y',
-    'TAA': '*',  // stop
-    'TAG': '*',  // stop
-    'CAT': 'H',
-    'CAC': 'H',
-    'CAA': 'Q',
-    'CAG': 'Q',
-    'AAT': 'N',
-    'AAC': 'N',
-    'AAA': 'K',
-    'AAG': 'K',
-    'GAT': 'D',
-    'GAC': 'D',
-    'GAA': 'E',
-    'GAG': 'E',
-    'TGT': 'C',
-    'TGC': 'C',
-    'TGA': '*',  // stop
-    'TGG': 'W',
-    'CGT': 'R',
-    'CGC': 'R',
-    'CGA': 'R',
-    'CGG': 'R',
-    'AGT': 'S',
-    'AGC': 'S',
-    'AGA': 'R',
-    'AGG': 'R',
-    'GGT': 'G',
-    'GGC': 'G',
-    'GGA': 'G',
-    'GGG': 'G'
-}
-
-function resolveUrlToPage(rel) {
-    return makeElement('a', null, {href: rel}).href;
-}
-
-//
-// Missing APIs
-// 
-
-if (!('trim' in String.prototype)) {
-    String.prototype.trim = function() {
-        return this.replace(/^\s+/, '').replace(/\s+$/, '');
-    };
-}
-
-if (typeof(module) !== 'undefined') {
-    module.exports = {
-        textXHR: textXHR,
-        relativeURL: relativeURL,
-        resolveUrlToPage: resolveUrlToPage,
-        shallowCopy: shallowCopy,
-        pusho: pusho,
-        pushnew: pushnew,
-        pushnewo: pushnewo,
-        arrayIndexOf: arrayIndexOf,
-        pick: pick,
-
-        makeElement: makeElement,
-        makeElementNS: makeElementNS,
-        removeChildren: removeChildren,
-
-        miniJSONify: miniJSONify,
-
-        Observed: Observed,
-        Awaited: Awaited,
-
-        AMINO_ACID_TRANSLATION: AMINO_ACID_TRANSLATION
-    }
-}
-});
